+++ symengine/utilities/matchpycpp/utils.h
